separate test deployment version into unit and non-unit tests
